Cleaning up because sciensano doesn't
Now doing a head request to see if the file exists because sciensano just returns a 200, even if the file doesn't exist. And I think they started doing this for future files as well, so the github action would ask for the file before it existed, sciensano would say it's fine, and kronacheck would have a bogus csv file.

So now I do the request and I did a

      grep -l "Not Found" data/* | xargs rm

To remove old bogus data